real/requirements.txt

pytesseract
Pillow
